@@ -14,10 +14,10 @@ classifier =
         Intended Audience :: Science/Research
 	License :: OSI Approved :: Apache Software License
         Operating System :: POSIX
-	Programming Language :: Python :: 3.8
 	Programming Language :: Python :: 3.9
 	Programming Language :: Python :: 3.10
 	Programming Language :: Python :: 3.11
+	Programming Language :: Python :: 3.12
 	Topic :: Scientific/Engineering :: Image Recognition
 	Topic :: Scientific/Engineering :: Artificial Intelligence
 
@@ -38,30 +38,31 @@ max_line_length = 160
 exclude = tests/*
 
 [options]
-python_requires = >=3.8,<=3.11.99
+python_requires = >=3.9,<3.13
 install_requires =
     jsonschema
     lxml
     requests
     click>=8.1
-    numpy~=1.23.0
+    numpy~=2.0.0
     Pillow>=9.2.0
     regex
-    scipy~=1.10.0 # bump up to latest release with py3.8 EOL
+    scipy~=1.13.0
     protobuf>=3.0.0
-    coremltools~=6.0
+    coremltools~=8.1
     jinja2~=3.0
-    python-bidi
+    python-bidi~=0.6.0
     torchvision>=0.5.0
-    torch~=2.1.0
-    scikit-learn~=1.2.1
-    scikit-image~=0.21.0
-    shapely~=1.8.5
+    torch~=2.4.0
+    scikit-learn~=1.5.0
+    scikit-image~=0.24.0
+    shapely>=2.0.6,~=2.0.6
     pyarrow
-    lightning~=2.2.0
+    htrmopo>=0.3,~=0.3
+    lightning~=2.4.0
     torchmetrics>=1.1.0
-    threadpoolctl~=3.4.0
-    importlib-resources>=1.3.0
+    threadpoolctl~=3.5.0
+    platformdirs
     rich
 
 [options.extras_require]
