@@ -28,6 +28,8 @@ from PIL import Image
 
 from kraken.ketos.util import (_expand_gt, _validate_manifests, message,
                                to_ptl_device)
+
+from kraken.lib.register import OPTIMIZERS, SCHEDULERS, STOPPERS, PRECISIONS
 from kraken.lib.default_specs import (SEGMENTATION_HYPER_PARAMS,
                                       SEGMENTATION_SPEC)
 from kraken.lib.exceptions import KrakenInputException
@@ -82,8 +84,7 @@ def _validate_merging(ctx, param, value):
               '--quit',
               show_default=True,
               default=SEGMENTATION_HYPER_PARAMS['quit'],
-              type=click.Choice(['early',
-                                 'fixed']),
+              type=click.Choice(STOPPERS),
               help='Stop condition for training. Set to `early` for early stopping or `fixed` for fixed number of epochs')
 @click.option('-N',
               '--epochs',
@@ -106,16 +107,13 @@ def _validate_merging(ctx, param, value):
 @click.option('-d', '--device', show_default=True, default='cpu', help='Select device to use (cpu, cuda:0, cuda:1, ...)')
 @click.option('--precision',
               show_default=True,
-              default='32',
-              type=click.Choice(['64', '32', 'bf16', '16']),
+              default='32-true',
+              type=click.Choice(PRECISIONS),
               help='Numerical precision to use for training. Default is 32-bit single-point precision.')
 @click.option('--optimizer',
               show_default=True,
               default=SEGMENTATION_HYPER_PARAMS['optimizer'],
-              type=click.Choice(['Adam',
-                                 'SGD',
-                                 'RMSprop',
-                                 'Lamb']),
+              type=click.Choice(OPTIMIZERS),
               help='Select optimizer')
 @click.option('-r', '--lrate', show_default=True, default=SEGMENTATION_HYPER_PARAMS['lrate'], help='Learning rate')
 @click.option('-m', '--momentum', show_default=True, default=SEGMENTATION_HYPER_PARAMS['momentum'], help='Momentum')
@@ -125,12 +123,7 @@ def _validate_merging(ctx, param, value):
               default=SEGMENTATION_HYPER_PARAMS['warmup'], help='Number of steps to ramp up to `lrate` initial learning rate.')
 @click.option('--schedule',
               show_default=True,
-              type=click.Choice(['constant',
-                                 '1cycle',
-                                 'exponential',
-                                 'cosine',
-                                 'step',
-                                 'reduceonplateau']),
+              type=click.Choice(SCHEDULERS),
               default=SEGMENTATION_HYPER_PARAMS['schedule'],
               help='Set learning rate scheduler. For 1cycle, cycle length is determined by the `--step-size` option.')
 @click.option('-g',
@@ -163,7 +156,7 @@ def _validate_merging(ctx, param, value):
 @click.option('-e', '--evaluation-files', show_default=True, default=None, multiple=True,
               callback=_validate_manifests, type=click.File(mode='r', lazy=True),
               help='File(s) with paths to evaluation data. Overrides the `-p` parameter')
-@click.option('--workers', show_default=True, default=1, type=click.IntRange(0), help='Number of worker proesses.')
+@click.option('--workers', show_default=True, default=1, type=click.IntRange(0), help='Number of data loading worker processes.')
 @click.option('--threads', show_default=True, default=1, type=click.IntRange(1), help='Maximum size of OpenMP/BLAS thread pool.')
 @click.option('--load-hyper-parameters/--no-load-hyper-parameters', show_default=True, default=False,
               help='When loading an existing model, retrieve hyper-parameters from the model')
