@@ -157,8 +157,8 @@ def serialize(results: 'Segmentation',
             reg = reg_dict[record.regions[0]]
             regs_with_lines.add(reg.id)
             region = {'id': reg.id,
-                      'bbox': max_bbox([reg.boundary]),
-                      'boundary': [list(x) for x in reg.boundary],
+                      'bbox': max_bbox([reg.boundary]) if reg.boundary else [],
+                      'boundary': [list(x) for x in reg.boundary] if reg.boundary else [],
                       'tags': reg.tags,
                       'lines': [],
                       'type': 'region'
@@ -256,6 +256,7 @@ def render_report(model: str,
                   chars: int,
                   errors: int,
                   char_accuracy: float,
+                  char_CI_accucary: float, #Case insensitive
                   word_accuracy: float,
                   char_confusions: 'Counter',
                   scripts: 'Counter',
@@ -286,6 +287,7 @@ def render_report(model: str,
               'chars': chars,
               'errors': errors,
               'character_accuracy': char_accuracy * 100,
+              'character_CI_accucary': char_CI_accucary * 100,
               'word_accuracy': word_accuracy * 100,
               'insertions': sum(insertions.values()),
               'deletions': deletions,
